Introduce compiler plugin level constants
