+++ grpc-compiler-plugin-tests/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     testImplementation group: 'org.ballerinalang', name: 'ballerina-tools-api', version: "${ballerinaLangVersion}"
     testImplementation group: 'org.ballerinalang', name: 'ballerina-parser', version: "${ballerinaLangVersion}"
     testImplementation project(":grpc-native")
+    testImplementation project(":grpc-compiler-plugin")
     testImplementation group: 'org.testng', name: 'testng', version: "${testngVersion}"
 }
